+++ src/hwenc_vpl/vpl_utils.h
@@ -18,11 +18,11 @@
 namespace sora {
 
 static mfxU32 ToMfxCodec(webrtc::VideoCodecType codec) {
-  return codec == webrtc::kVideoCodecVP8    ? MFX_CODEC_VP8
-         : codec == webrtc::kVideoCodecVP9  ? MFX_CODEC_VP9
-         : codec == webrtc::kVideoCodecH264 ? MFX_CODEC_AVC
-         : codec == webrtc::kVideoCodecH265 ? MFX_CODEC_HEVC
-                                            : MFX_CODEC_AV1;
+  return codec == webrtc::kVideoCodecVP8    ? (mfxU32)MFX_CODEC_VP8
+         : codec == webrtc::kVideoCodecVP9  ? (mfxU32)MFX_CODEC_VP9
+         : codec == webrtc::kVideoCodecH264 ? (mfxU32)MFX_CODEC_AVC
+         : codec == webrtc::kVideoCodecH265 ? (mfxU32)MFX_CODEC_HEVC
+                                            : (mfxU32)MFX_CODEC_AV1;
 }
 
 static std::string CodecToString(mfxU32 codec) {