+++ src/sora_signaling.cpp
@@ -94,7 +94,7 @@ void SoraSignaling::Disconnect() {
       return;
     }
 
-    self->DoInternalDisconnect(boost::none, "", "");
+    self->DoInternalDisconnect(std::nullopt, "", "");
   });
 }
 
@@ -332,9 +332,6 @@ void SoraSignaling::DoSendConnect(bool redirect) {
 
   if (!config_.audio) {
     m["audio"] = false;
-  } else if (config_.audio && config_.audio_codec_type.empty() &&
-             config_.audio_bit_rate == 0) {
-    m["audio"] = true;
   } else {
     m["audio"] = boost::json::object();
     if (!config_.audio_codec_type.empty()) {
@@ -342,6 +339,14 @@ void SoraSignaling::DoSendConnect(bool redirect) {
     }
     if (config_.audio_bit_rate != 0) {
       m["audio"].as_object()["bit_rate"] = config_.audio_bit_rate;
+    }
+    if (!config_.audio_opus_params.is_null()) {
+      m["audio"].as_object()["opus_params"] = config_.audio_opus_params;
+    }
+
+    // オプションの設定が行われてなければ単に true を設定
+    if (m["audio"].as_object().empty()) {
+      m["audio"] = true;
     }
   }
 
@@ -377,14 +382,22 @@ void SoraSignaling::DoSendConnect(bool redirect) {
       if (d.compress) {
         obj["compress"] = *d.compress;
       }
+      if (d.header) {
+        obj["header"] = boost::json::array(d.header->begin(), d.header->end());
+      }
       ar.push_back(obj);
     }
     m["data_channels"] = ar;
   }
 
-  if (config_.forwarding_filter) {
+  auto forwarding_filter_to_json = [](const SoraSignalingConfig::ForwardingFilter& f) -> boost::json::value {
     boost::json::object obj;
-    auto& f = *config_.forwarding_filter;
+    if (f.name) {
+      obj["name"] = *f.name;
+    }
+    if (f.priority) {
+      obj["priority"] = *f.priority;
+    }
     if (f.action) {
       obj["action"] = *f.action;
     }
@@ -409,7 +422,19 @@ void SoraSignaling::DoSendConnect(bool redirect) {
     if (f.metadata) {
       obj["metadata"] = *f.metadata;
     }
-    m["forwarding_filter"] = obj;
+    return obj;
+  };
+
+  if (config_.forwarding_filter) {
+    m["forwarding_filter"] = forwarding_filter_to_json(*config_.forwarding_filter);
+  }
+
+  if (config_.forwarding_filters) {
+    boost::json::array ar;
+    for (const auto& f : *config_.forwarding_filters) {
+      ar.push_back(forwarding_filter_to_json(f));
+    }
+    m["forwarding_filters"] = ar;
   }
 
   std::string text = boost::json::serialize(m);
@@ -565,7 +590,7 @@ SoraSignaling::CreatePeerConnection(boost::json::value jconfig) {
 }
 
 void SoraSignaling::DoInternalDisconnect(
-    boost::optional<SoraSignalingErrorCode> force_error_code,
+    std::optional<SoraSignalingErrorCode> force_error_code,
     std::string reason,
     std::string message) {
   assert(state_ == State::Connected);
@@ -576,7 +601,7 @@ void SoraSignaling::DoInternalDisconnect(
                    root_message = message](bool succeeded,
                                            SoraSignalingErrorCode error_code,
                                            std::string message) {
-    if (force_error_code == boost::none) {
+    if (force_error_code == std::nullopt) {
       self->SendOnDisconnect(error_code, message);
     } else {
       self->SendOnDisconnect(*force_error_code,
@@ -605,7 +630,7 @@ void SoraSignaling::DoInternalDisconnect(
 
   if (using_datachannel_ && ws_connected_) {
     std::string text =
-        force_error_code == boost::none
+        force_error_code == std::nullopt
             ? R"({"type":"disconnect","reason":"NO-ERROR"})"
             : R"({"type":"disconnect","reason":")" + reason + "\"}";
     webrtc::DataBuffer disconnect = ConvertToDataBuffer("signaling", text);
@@ -751,10 +776,6 @@ void SoraSignaling::OnConnect(boost::system::error_code ec,
                      [ws](std::shared_ptr<Websocket> p) { return p == ws; }),
       connecting_wss_.end());
 
-  if (state_ == State::Closed) {
-    return;
-  }
-
   if (ec) {
     RTC_LOG(LS_WARNING) << "Failed Websocket handshake: " << ec
                         << " url=" << url << " state=" << (int)state_
@@ -771,9 +792,10 @@ void SoraSignaling::OnConnect(boost::system::error_code ec,
     return;
   }
 
-  if (state_ == State::Connected || state_ == State::Redirecting) {
+  if (state_ == State::Connected || state_ == State::Redirecting ||
+      state_ == State::Closing || state_ == State::Closed) {
     // 既に他の接続が先に完了していたので、切断する
-    ws->Close([self = shared_from_this(), ws](boost::system::error_code) {},
+    ws->Close([ws](boost::system::error_code) {},
               config_.websocket_close_timeout);
     return;
   }
@@ -1214,7 +1236,7 @@ void SoraSignaling::DoConnect() {
     } else {
       ws.reset(new Websocket(*config_.io_context));
     }
-    if (config_.user_agent != boost::none) {
+    if (config_.user_agent != std::nullopt) {
       ws->SetUserAgent(*config_.user_agent);
     }
     ws->Connect(url, std::bind(&SoraSignaling::OnConnect, shared_from_this(),
@@ -1595,7 +1617,7 @@ void SoraSignaling::OnMessage(
     return;
   }
 
-  boost::json::error_code ec;
+  boost::system::error_code ec;
   auto json = boost::json::parse(data, ec);
   if (ec) {
     RTC_LOG(LS_ERROR) << "JSON Parse Error ec=" << ec.message();
@@ -1644,6 +1666,24 @@ void SoraSignaling::OnMessage(
           },
           CreateIceError(
               "Failed to SetOffer in re-offer message via DataChannel"));
+    } else if (type == "close") {
+      // グレースフルシャットダウンする
+      int code = json.at("code").to_number<int>();
+      std::string reason = json.at("reason").as_string().c_str();
+      // on_close を設定しておいて、あとは DataChannel が閉じるのを待つだけ
+      dc_->SetOnClose([self = shared_from_this(), code,
+                       reason](boost::system::error_code ec) {
+        if (code == 1000) {
+          self->SendOnDisconnect(
+              SoraSignalingErrorCode::CLOSE_SUCCEEDED,
+              "Succeeded to close DataChannel: reason=" + reason);
+        } else {
+          self->SendOnDisconnect(
+              SoraSignalingErrorCode::CLOSE_SUCCEEDED,
+              "Failed to close DataChannel: code=" + std::to_string(code) +
+                  " ec=" + ec.message());
+        }
+      });
     }
     return;
   }